+++ scripts/macros/2024/speciesFeatures/aasimar/healingHands.js
@@ -1,0 +1,5 @@
+export let healingHands = {
+    name: 'Healing Hands',
+    version: '1.3.146',
+    rules: 'modern'
+};